web/client: refactored entity forms
