+++ web/client/src/entities/CallCsvScheduler/CallCsvScheduler.tsx
@@ -2,7 +2,8 @@ import SettingsApplications from '@mui/icons-material/SettingsApplications';
 import EntityInterface from 'lib/entities/EntityInterface';
 import _ from 'lib/services/translations/translate';
 import defaultEntityBehavior, { MarshallerValues } from 'lib/entities/DefaultEntityBehavior';
-import Form, { foreignKeyGetter } from './Form';
+import Form from './Form';
+import { foreignKeyGetter } from './useFkChoices';
 import LastExecution from './Field/LastExecution';
 import { CallCsvSchedulerProperties } from './CallCsvSchedulerProperties';
 import { PartialPropertyList } from 'lib/services/api/ParsedApiSpecInterface';
