+++ web/client/src/entities/CompanyService/useFkChoices.tsx
@@ -1,0 +1,64 @@
+import { FkChoices } from 'lib/entities/DefaultEntityBehavior';
+import { useEffect, useState } from 'react';
+import ServiceSelectOptions from 'entities/Service/SelectOptions';
+import { CompanyServicePropertyList } from './CompanyServiceProperties';
+import axios, { CancelToken } from 'axios';
+import { ForeignKeyGetterType } from 'lib/entities/EntityInterface';
+
+export const foreignKeyGetter: ForeignKeyGetterType = async (token?: CancelToken): Promise<any> => {
+
+    const response: CompanyServicePropertyList<unknown> = {};
+    const promises: Array<Promise<unknown>> = [];
+
+    promises[promises.length] = ServiceSelectOptions(
+        (options: any) => {
+            response.service = options;
+        },
+        undefined,
+        token
+    );
+
+    await Promise.all(promises);
+
+    return response;
+};
+
+const useFkChoices = (): FkChoices => {
+
+    const [fkChoices, setFkChoices] = useState<FkChoices>({});
+
+    useEffect(
+        () => {
+
+            let mounted = true;
+
+            const CancelToken = axios.CancelToken;
+            const source = CancelToken.source();
+
+            foreignKeyGetter(source.token).then((options) => {
+
+                if (!mounted) {
+                    return;
+                }
+
+                setFkChoices((fkChoices: any) => {
+                    return {
+                        ...fkChoices,
+                        ...options
+                    }
+                });
+            });
+
+
+            return () => {
+                mounted = false;
+                source.cancel();
+            }
+        },
+        []
+    );
+
+    return fkChoices;
+}
+
+export default useFkChoices;
