+++ web/client/src/entities/Fax/useFkChoices.tsx
@@ -1,0 +1,62 @@
+import { FkChoices } from 'lib/entities/DefaultEntityBehavior';
+import { useEffect, useState } from 'react';
+import DdiSelectOptions from 'entities/Ddi/SelectOptions';
+import { FaxPropertyList } from './FaxProperties';
+import axios, { CancelToken } from 'axios';
+import { ForeignKeyGetterType } from 'lib/entities/EntityInterface';
+
+export const foreignKeyGetter: ForeignKeyGetterType = async (token?: CancelToken): Promise<any> => {
+
+    const response: FaxPropertyList<Array<string | number>> = {};
+    const promises: Array<Promise<unknown>> = [];
+
+    promises[promises.length] = DdiSelectOptions(
+        (options: any) => {
+            response.outgoingDdi = options;
+        },
+        token
+    );
+
+    await Promise.all(promises);
+
+    return response;
+};
+
+const useFkChoices = (): FkChoices => {
+
+    const [fkChoices, setFkChoices] = useState<FkChoices>({});
+
+    useEffect(
+        () => {
+
+            let mounted = true;
+
+            const CancelToken = axios.CancelToken;
+            const source = CancelToken.source();
+
+            foreignKeyGetter(source.token).then((options) => {
+
+                if (!mounted) {
+                    return;
+                }
+
+                setFkChoices((fkChoices: any) => {
+                    return {
+                        ...fkChoices,
+                        ...options
+                    }
+                });
+            });
+
+            return () => {
+                mounted = false;
+                source.cancel();
+            }
+        },
+        []
+    );
+
+    return fkChoices;
+}
+
+export default useFkChoices;
