+++ web/client/src/entities/HuntGroup/useFkChoices.tsx
@@ -1,0 +1,87 @@
+import { FkChoices } from 'lib/entities/DefaultEntityBehavior';
+import { useEffect, useState } from 'react';
+import LocutionSelectOptions from 'entities/Locution/SelectOptions';
+import CountrySelectOptions from 'entities/Country/SelectOptions';
+import ExtensionSelectOptions from 'entities/Extension/SelectOptions';
+import UserSelectOptions from 'entities/User/SelectOptions';
+import { HuntGroupPropertyList } from './HuntGroupProperties';
+import { ForeignKeyGetterType } from 'lib/entities/EntityInterface';
+import axios, { CancelToken } from 'axios';
+
+export const foreignKeyGetter: ForeignKeyGetterType = async (token?: CancelToken): Promise<any> => {
+
+    const response: HuntGroupPropertyList<Array<string | number>> = {};
+    const promises: Array<Promise<unknown>> = [];
+
+    promises[promises.length] = LocutionSelectOptions(
+        (options: any) => {
+            response.noAnswerLocution = options;
+        },
+        token
+    );
+
+    promises[promises.length] = CountrySelectOptions(
+        (options: any) => {
+            response.noAnswerNumberCountry = options;
+        },
+        token
+    );
+
+    promises[promises.length] = ExtensionSelectOptions(
+        (options: any) => {
+            response.noAnswerExtension = options;
+        },
+        token
+    );
+
+    promises[promises.length] = UserSelectOptions(
+        (options: any) => {
+            response.noAnswerVoiceMailUser = options;
+        },
+        token
+    );
+
+    await Promise.all(promises);
+
+    return response;
+};
+
+
+const useFkChoices = (): FkChoices => {
+
+    const [fkChoices, setFkChoices] = useState<FkChoices>({});
+
+    useEffect(
+        () => {
+
+            let mounted = true;
+
+            const CancelToken = axios.CancelToken;
+            const source = CancelToken.source();
+
+            foreignKeyGetter(source.token).then((options) => {
+
+                if (!mounted) {
+                    return;
+                }
+
+                setFkChoices((fkChoices: any) => {
+                    return {
+                        ...fkChoices,
+                        ...options
+                    }
+                });
+            });
+
+            return () => {
+                mounted = false;
+                source.cancel();
+            }
+        },
+        []
+    );
+
+    return fkChoices;
+}
+
+export default useFkChoices;
