+++ web/client/src/entities/Ivr/useFkChoices.tsx
@@ -1,0 +1,93 @@
+import { FkChoices } from 'lib/entities/DefaultEntityBehavior';
+import { useEffect, useState } from 'react';
+import CountrySelectOptions from 'entities/Country/SelectOptions';
+import LocutionSelectOptions from 'entities/Locution/SelectOptions';
+import ExtensionSelectOptions from 'entities/Extension/SelectOptions';
+import UserSelectOptions from 'entities/User/SelectOptions';
+import { IvrPropertyList } from './IvrProperties';
+import axios, { CancelToken } from 'axios';
+import { ForeignKeyGetterType } from 'lib/entities/EntityInterface';
+
+export const foreignKeyGetter: ForeignKeyGetterType = async (token?: CancelToken): Promise<any> => {
+
+    const response: IvrPropertyList<Array<string | number>> = {};
+    const promises: Array<Promise<unknown>> = [];
+
+    promises[promises.length] = LocutionSelectOptions(
+        (options: any) => {
+            response.welcomeLocution = options;
+            response.noInputLocution = options;
+            response.errorLocution = options;
+            response.successLocution = options;
+        },
+        token
+    );
+
+    promises[promises.length] = CountrySelectOptions(
+        (options: any) => {
+            response.noInputNumberCountry = options;
+            response.errorNumberCountry = options;
+        },
+        token
+    );
+
+    promises[promises.length] = ExtensionSelectOptions(
+        (options: any) => {
+            response.noInputExtension = options;
+            response.errorExtension = options;
+            response.excludedExtensionIds = options;
+        },
+        token
+    );
+
+    promises[promises.length] = UserSelectOptions(
+        (options: any) => {
+            response.noInputVoiceMailUser = options;
+            response.errorVoiceMailUser = options;
+        },
+        token
+    );
+
+    await Promise.all(promises);
+
+    return response;
+};
+
+const useFkChoices = (): FkChoices => {
+
+    const [fkChoices, setFkChoices] = useState<FkChoices>({});
+
+    useEffect(
+        () => {
+
+            let mounted = true;
+
+            const CancelToken = axios.CancelToken;
+            const source = CancelToken.source();
+
+            foreignKeyGetter(source.token).then((options) => {
+
+                if (!mounted) {
+                    return;
+                }
+
+                setFkChoices((fkChoices: any) => {
+                    return {
+                        ...fkChoices,
+                        ...options
+                    }
+                });
+            });
+
+            return () => {
+                mounted = false;
+                source.cancel();
+            }
+        },
+        []
+    );
+
+    return fkChoices;
+}
+
+export default useFkChoices;
