+++ web/client/src/entities/OutgoingDdiRule/OutgoingDdiRule.tsx
@@ -2,7 +2,8 @@ import SettingsApplications from '@mui/icons-material/SettingsApplications';
 import EntityInterface from 'lib/entities/EntityInterface';
 import _ from 'lib/services/translations/translate';
 import defaultEntityBehavior from 'lib/entities/DefaultEntityBehavior';
-import Form, { foreignKeyGetter } from './Form'
+import Form from './Form';
+import { foreignKeyGetter } from './useFkChoices'
 import { EntityValues } from 'lib/services/entity/EntityService';
 import entities from '../index';
 import genericForeignKeyResolver from 'lib/services/api/genericForeigKeyResolver';
@@ -33,7 +34,6 @@ const properties: PartialPropertyList = {
         label: _('Forced DDI'),
     }
 };
-
 
 async function foreignKeyResolver(data: EntityValues): Promise<EntityValues> {
