+++ web/client/src/entities/RatingProfile/RatingProfile.tsx
@@ -2,7 +2,8 @@ import SettingsApplications from '@mui/icons-material/SettingsApplications';
 import EntityInterface from 'lib/entities/EntityInterface';
 import _ from 'lib/services/translations/translate';
 import defaultEntityBehavior from 'lib/entities/DefaultEntityBehavior';
-import Form, { foreignKeyGetter } from './Form'
+import Form from './Form';
+import { foreignKeyGetter } from './useFkChoices';
 import RatingPlanGroup from '../RatingPlanGroup/RatingPlanGroup';
 import entities from '../index';
 import genericForeignKeyResolver from 'lib/services/api/genericForeigKeyResolver';
