+++ web/client/src/entities/Terminal/Terminal.tsx
@@ -3,7 +3,8 @@ import EntityInterface from 'lib/entities/EntityInterface';
 import defaultEntityBehavior from 'lib/entities/DefaultEntityBehavior';
 import genericForeignKeyResolver from 'lib/services/api/genericForeigKeyResolver';
 import _ from 'lib/services/translations/translate';
-import Form, { foreignKeyGetter } from './Form'
+import Form from './Form';
+import { foreignKeyGetter } from './useFkChoices';
 import entities from '../index';
 import { TerminalProperties, TerminalPropertiesList } from './TerminalProperties';
