+++ web/client/src/entities/User/User.tsx
@@ -2,7 +2,8 @@ import SettingsApplications from '@mui/icons-material/SettingsApplications';
 import EntityInterface from 'lib/entities/EntityInterface';
 import _ from 'lib/services/translations/translate';
 import defaultEntityBehavior from 'lib/entities/DefaultEntityBehavior';
-import Form, { foreignKeyGetter } from './Form'
+import Form from './Form'
+import { foreignKeyGetter } from './useFkChoices'
 import genericForeignKeyResolver from 'lib/services/api/genericForeigKeyResolver';
 import entities from '../index';
 import { UserProperties, UserPropertiesList } from './UserProperties';
